+++ src-gen/main/java/com/movisens/smartgattlib/attributes/ManufacturerNameString.java
@@ -4,8 +4,9 @@ import com.movisens.smartgattlib.Characteristics;
 import com.movisens.smartgattlib.helper.AbstractReadAttribute;
 import com.movisens.smartgattlib.helper.Characteristic;
 import com.movisens.smartgattlib.helper.GattByteBuffer;
+import com.movisens.smartgattlib.helper.PlainTextAttribute;
 
-public class ManufacturerNameString extends AbstractReadAttribute
+public class ManufacturerNameString extends AbstractReadAttribute implements PlainTextAttribute
 {
 
 	public static final Characteristic<ManufacturerNameString> CHARACTERISTIC = Characteristics.MANUFACTURER_NAME_STRING;