+++ src/main/java/com/movisens/smartgattlib/helper/Characteristic.java
@@ -1,11 +1,17 @@
 package com.movisens.smartgattlib.helper;
 
+import java.lang.reflect.Constructor;
+import java.util.Arrays;
+import java.util.function.Predicate;
+
 import com.movisens.smartgattlib.attributes.DefaultAttribute;
+import com.movisens.smartgattlib.security.CryptoManagerProvider;
 
 public class Characteristic<T extends AbstractAttribute> extends UuidObject
 {
+
     private Class<T> attributeClass;
-    
+
     Characteristic<?> requiredCharacteristics[];
 
     public Characteristic<?>[] getRequiredCharacteristics()
@@ -20,27 +26,50 @@ public class Characteristic<T extends AbstractAttribute> extends UuidObject
         this.requiredCharacteristics = requiredCharacteristics;
     }
 
-    public AbstractAttribute createAttribute(byte[] data)
+    public AbstractAttribute createAttribute(byte[] incommingData)
     {
         try
         {
-            return attributeClass.getConstructor(byte[].class).newInstance(data);
+            Constructor<T> constructor = attributeClass.getConstructor(byte[].class);
+
+            if (isEncryptionAllowed())
+            {
+                return constructor.newInstance(CryptoManagerProvider.get().processAfterReceive(incommingData));
+            }
+            else
+            {
+                return constructor.newInstance(incommingData);
+            }
         }
         catch (Throwable e)
         {
             e.printStackTrace();
-            return new DefaultAttribute(data);
+            return new DefaultAttribute(incommingData);
         }
     }
-    
+
     public Class<T> getAttributeClass()
     {
         return attributeClass;
     }
-    
+
+    public boolean isEncryptionAllowed()
+    {
+        return Arrays.stream(attributeClass.getInterfaces()).filter(new Predicate<Class<?>>()
+        {
+
+            @Override
+            public boolean test(Class<?> i)
+            {
+                return i.getName().equals(PlainTextAttribute.class.getName());
+            }
+        }).count() == 0;
+    }
+
     @Override
     public String toString()
     {
         return getName();
     }
+
 }