+++ src/main/java/com/movisens/smartgattlib/helper/PlainTextAttribute.java
@@ -1,0 +1,11 @@
+package com.movisens.smartgattlib.helper;
+
+/**
+ * Attributes that implements the PlainTextAttribute interface will not be encrypted.
+ * 
+ * @author ulrich.grossmann
+ *
+ */
+public interface PlainTextAttribute
+{
+}