+++ src/main/java/com/movisens/smartgattlib/security/CryptoManagerProvider.java
@@ -1,0 +1,18 @@
+package com.movisens.smartgattlib.security;
+
+
+public class CryptoManagerProvider
+{
+    private static CryptoManager cryptoManager = new CryptoManager();
+
+    public static CryptoManager get()
+    {
+        return CryptoManagerProvider.cryptoManager;
+    }
+
+    public static void setSecurityManager(CryptoManager cryptoManager)
+    {
+        CryptoManagerProvider.cryptoManager = cryptoManager;
+    }
+    
+}